requirements.txt
fastapi
uvicorn
python-docx
pydantic
jinja2
sqlalchemy
aiofiles
python-multipart